feat(Github actions): modify ready to merge action to use a simpler github script
