--- libs/shared/gh-actions/tsconfig.app.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "module": "commonjs",
-    "types": ["node"]
-  },
-  "include": ["src/**/*.ts"]
-}
