--- libs/shared/gh-actions/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "files": [],
-  "include": [],
-  "references": [
-    {
-      "path": "./tsconfig.app.json"
-    }
-  ],
-  "compilerOptions": {
-    "esModuleInterop": true
-  }
-}
